+++ src/javascript/app_2/App/Components/Layout/header.jsx
@@ -2,9 +2,8 @@ import PropTypes           from 'prop-types';
 import React               from 'react';
 import { AccountBalance }  from '../Elements/account_balance.jsx';
 import AccountSwitcher     from '../Elements/account_switcher.jsx';
-import {
-    ToggleDrawer,
-    DrawerItem }           from '../Elements/Drawer/index.jsx';
+import { Notifications }   from '../Elements/Notifications/notifications.jsx';
+import { ToggleDrawer }    from '../Elements/Drawer/index.jsx';
 import { MenuDrawer }      from '../Elements/menu_drawer.jsx';
 import { BinaryLink }      from '../../routes';
 import Url                 from '../../../../_common/url';
@@ -43,11 +42,7 @@ class Header extends React.Component {
                             alignment='right'
                             icon_link={Url.urlForStatic('images/app_2/header/icons/ic_notification_light.svg')}
                         >
-                            {placeholder_alerts_list.length ?
-                                <Notifications list={placeholder_alerts_list} />
-                              :
-                                <NoNotifications />
-                            }
+                            <Notifications />
                         </ToggleDrawer>
                     </div>
                 </header>
@@ -55,37 +50,6 @@ class Header extends React.Component {
         );
     }
 }
-
-const placeholder_alerts_list = {
-    'alert1': 'message',
-    'alert2': 'message',
-    'alert3': 'message',
-    'alart4': 'message',
-};
-
-const Notifications = ({ list }) => (
-    list.map((item, idx) => (
-        <React.Fragment key={idx}>
-            <DrawerItem text={item[idx]} />
-        </React.Fragment>
-    ))
-);
-
-const NoNotifications = () => (
-    <div className='no-notifications-container'>
-        <div>
-            <img src={Url.urlForStatic('images/app_2/header/icons/ic_notification_light.svg')} />
-        </div>
-        <div>
-            <h4>No Notifications</h4>
-            <span className='no-notifications_message'>You have yet to receive any notifications</span>
-        </div>
-    </div>
-);
-
-Notifications.propTypes = {
-    'list': PropTypes.object,
-};
 
 Header.propTypes = {
     items: PropTypes.array,
